Added UIView Helper Methods to detect whether a view is currently part of the view hierarchy
I tried to add some documentation as well. I wrote this for nRF Connect, but then thought it was "open" enough to be added to the common library. I don't think it'll be used much since modern apps now just rely on SwiftUI to handle the magic for events such as rotation, resizing and so on. This is where this helper is needed for nRF Connect.