Ported "CancellableTextField" from nRF-Connect
I didn't even remember we wrote this! I found it trying to copy something else from nRF Connect, and noticed the TextField looked custom and also very good. To my surprise, of course. But of course, most of it is copied, or based on a copy from the web. No AI.