Ported a modified version of nRF Connect's "DataParser"
This is, so that we can use a custom View that nRF Connect uses in Common, so it's useable in other apps. But it has this "plumbing" component, which we want to keep extensible and private at the moment. Not really sure about what's the best way to go around that, unfortunately.