Ported nRF Connect's NordicDataField
This was written with the Peripheral re-write, which added a lot more SwiftUI code to nRF Connect. And this is the "DataField" UI we want to use for the future, so it makes sense to move it here, to make it reusable, and hence better across all of our other apps. Just like part of getting here, was the DataParser API, which got better with the enum.