Defense against GlucoseMeasurement Data Parsing Crash
nRF Connect attempts to parse anything as Data. So, bad "Data" might be given to parse, so we must be defensive and not think that just because we're getting called, we're going to get valid Data to parse. The goal is to not crash, and if gibberish is what we get, gibberish is what should go out. But no crashing.

This won't be the last of it, of course.