Added Imperfect blink() Modifier
Dear SwiftUI, why are you so glitchy and buggy? The non-deprecated API doesn't work... It either doesn't animate, or it animates at something like half rate for the effect and 10x times the duration of it. I guess I'm not in a mood to solve this today. I just want to get to other things I want to do at the moment, to be honest.