Re-raised Package version to iOS 15 / macOS 12
This is because, with package level below 15, the shaders in the package just don't compile. Which, sucks big time, honestly. But anyway, Apple is quietly forcing everything to be 15 at minimum. Xcode won't even build and deploy to anything lower than 15, so perhaps everything needs to be minimum 15 anyways but, let's keep nRF CDM and its attached libraries in 13.