Network is no longer always receiving on DispatchQueue.main
It doesn't affect anything when the user of the library works in this way. But if the user wants to use the Network from the background, like from an actor, then this entire API chain is kind of useless because it'll trigger data races and the like. If not crashes under Swift 6.