Removed last shader
We have ongoing limitations with this "Common" library and its dependencies due to the minimum version thing. Where, even if we tried to keep two versions, the shaders are not loaded above iOS 15 anyway. Someday we might be able to raise it to 15 and get shaders back. But that's not today.